Extract shared trustee test helper and strengthen E2E assertions
The new-tab navigation setup was duplicated across four accessibility
test files. Extract it into openFirstTrusteeProfileInNewTab in a new
trustee-common.ts peer file. Also strengthen the E2E navigation test
to assert that a new tab was opened, the original page stays on
/trustees, and the browser context page count increases by one.

Jira ticket: N/A
